requirements.txt
dropbox
pdfplumber
python-pptx
python-docx
pytesseract
opencv-python
pdf2image
Pillow
python-dotenv
langchain
langchain_community
langchain_openai
sentence-transformers
faiss-cpu
openai
fastapi
uvicorn